--- Settings.hs
@@ -1,7 +1,0 @@
--- length of grid
-gridLength:: Int
-gridLength = 25
-
--- tile's pixel height/width
-tileSize :: Int
-tileSize = 25
